pyproject.toml
[project]
name = "nonebot-plugin-bot-monitor"
version = "1.0.0"
description = "NoneBot2 插件：监控 Bot 在线状态，检测 Bot 离线并发送邮件通知"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "TogetsuDo", email = "togetsudo@outlook.com" }]
keywords = ["nonebot", "nonebot2", "bot-status", "monitor", "onebot"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Communications :: Chat",
]
dependencies = [
  "aiosmtplib>=5.0.0",
  "nonebot-adapter-onebot>=2.4.0,<3.0.0",
  "nonebot-plugin-apscheduler>=0.5.0,<1.0.0",
  "nonebot2>=2.4.0,<3.0.0",
]

[project.urls]
Homepage = "https://github.com/TogetsuDo/nonebot-plugin-bot-monitor"
Issues = "https://github.com/TogetsuDo/nonebot-plugin-bot-monitor/issues"
Repository = "https://github.com/TogetsuDo/nonebot-plugin-bot-monitor.git"

[dependency-groups]
dev = [
  "bump-my-version>=1.2.7",
  "nonebot2[fastapi]>=2.5.0,<3.0.0",
  "poethepoet>=0.42.1",
  "ruff>=0.15.2,<1.0.0",
  { include-group = "test" },
]
test = [
  "nonebot-adapter-onebot>=2.4.6,<3.0.0",
  "nonebot2[fastapi]>=2.5.0,<3.0.0",
  "nonebug>=0.4.4,<1.0.0",
  "pytest>=9.0.2,<10.0.0",
  "poethepoet>=0.42.1",
  "pytest-asyncio>=1.3.0,<1.4.0",
  "pytest-cov>=7.0.0",
  "pytest-xdist>=3.8.0,<4.0.0",
]

[build-system]
requires = ["uv_build>=0.10.0,<1.0.0"]
build-backend = "uv_build"


[tool.bumpversion]
current_version = "1.0.0"
commit = true
message = "release: bump version from {current_version} to {new_version}"
tag = true

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = "version = \"{current_version}\""
replace = "version = \"{new_version}\""

[[tool.bumpversion.files]]
filename = "uv.lock"
search = "name = \"nonebot-plugin-bot-monitor\"\nversion = \"{current_version}\""
replace = "name = \"nonebot-plugin-bot-monitor\"\nversion = \"{new_version}\""

[tool.coverage.report]
exclude_lines = [
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
  "@overload",
  "except ImportError:",
]

[tool.nonebot]
plugins = ["nonebot_plugin_bot_monitor"]

[tool.poe.tasks]
test = "pytest --cov=src --cov-report xml --junitxml=./junit.xml -n auto"
bump = "bump-my-version bump"
show-bump = "bump-my-version show-bump"

[tool.pyright]
pythonVersion = "3.12"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
executionEnvironments = [
  { root = "./tests", extraPaths = [
    "./src",
  ] },
  { root = "./src" },
]
typeCheckingMode = "standard"
disableBytesTypePromotions = true

[tool.pytest]
addopts = [
  "--import-mode=prepend",
  "--strict-markers",
  "--tb=short",
  "-ra",
  "-s",
  "-v",
]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.ruff]
line-length = 88

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
  "F",
  "W",
  "E",
  "I",
  "UP",
  "ASYNC",
  "C4",
  "T10",
  "T20",
  "PYI",
  "PT",
  "Q",
  "TID",
  "RUF",
]
ignore = [
  "E402",
  "UP037",
  "RUF001",
  "RUF002",
  "RUF003",
  "W191",
  "TID252",
]

[tool.ruff.lint.isort]
length-sort = true
known-first-party = ["tests/*"]
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true
